Change the call ABI for function from size_t argc to unsigned argc. Implement static booleans, do not create dynamic booleans at runtime.

+++ tinyclj/types/Object.h
@@ -24,7 +24,7 @@ struct Object;
 struct TCString;
 
 using CallFn = const Object *(*)(const Object *self,
-                                 size_t argc,
+                                 unsigned argc,
                                  const struct Object **argv);
 
 using UnaryFn = const Object *(*)(const Object *self);
